WIP : incrementally changing graph - Trying to understand Graph structure - Looking for ways to reduce mem consumption - Packing pin + edge into one word
