+++ include/POLite/Seq.h
@@ -24,23 +24,26 @@ template <class T> class Seq
     T* elems;
 
     // Constructors
-    Seq() { init(4096); }
+    Seq() { init(4096/sizeof(T)); }  // dt10 - original default of 4096 bytes creates a lot of memory pressure for large graphs
     Seq(int initialSize) { init(initialSize); }
 
     // Copy constructor
     Seq(const Seq<T>& seq) {
       init(seq.maxElems);
       numElems = seq.numElems;
-      for (int i = 0; i < seq.numElems; i++)
-        elems[i] = seq.elems[i];
+      std::copy(seq.elems, seq.elems+seq.numElems, elems);
+      //for (int i = 0; i < seq.numElems; i++)
+      //  elems[i] = seq.elems[i];
     }
 
     // Set capacity of sequence
     void setCapacity(int n) {
       maxElems = n;
       T* newElems = new T[maxElems];
-      for (int i = 0; i < numElems-1; i++)
-        newElems[i] = elems[i];
+      std::move(elems, elems+numElems, newElems);
+      // BUG : dt10 : Pretty sure the upper bound of numElems-1 is a bug?
+      //for (int i = 0; i < numElems-1; i++)
+      //  newElems[i] = elems[i];
       delete [] elems;
       elems = newElems;
     }
@@ -111,19 +114,35 @@ template <class T> class Seq
 
     // Remove an element from a sequence
     void remove(T x) {
-      for (int i = 0; i < numElems; i++)
+      for (int i = 0; i < numElems; i++){
         if (elems[i] == x) {
-          for (int j = i; j < numElems-1; j++)
+          std::move(elems+1,elems+numElems, elems);
+          // BUG : dt10. The 
+          /*for (int j = i; j < numElems-1; j++)
             elems[j] = elems[j+1];
+          */
           numElems--;
           return;
         }
+      }
     }
 
     // Destructor
     ~Seq()
     {
       delete [] elems;
+    }
+
+    const T &operator[](size_t i) const
+    {
+      assert(i<numElems);
+      return elems[i];
+    }
+
+    T &operator[](size_t i)
+    {
+      assert(i<numElems);
+      return elems[i];
     }
 };
