+++ include/config.h
@@ -1,0 +1,141 @@
+#ifndef _CONFIG_H_
+#define _CONFIG_H_
+#define TinselFPDivLatency 14
+#define TinselCoresPerFPU 4
+#define TinselLogLinesPerMem 27
+#define TinselFetcherLogBeatBufferSize 5
+#define TinselWordsPerSRAMBeat 2
+#define TinselLogBytesPerDRAM 31
+#define TinselLogCoresPerBoard 6
+#define TinselWordsPerBeat 8
+#define TinselLogReliableLinkRecvBufferSize 9
+#define TinselLogDRAMsPerBoard 1
+#define TinselLogTransmitBufferLen 2
+#define TinselFPConvertLatency 6
+#define TinselInstrsPerCore 2048
+#define TinselMeshXBits1 4
+#define TinselMeshXBitsWithinBox 2
+#define TinselBytesPerDRAM 2147483648
+#define TinselFPCompareLatency 3
+#define TinselFPAddSubLatency 14
+#define TinselLogBytesPerFlit 4
+#define TinselLogFetcherFlitBufferSize 5
+#define TinselLogEastWestLinks 0
+#define TinselNumEastWestLinks 1
+#define TinselBytesPerSRAMBeat 8
+#define TinselMailboxMeshXLen 4
+#define TinselLinkTimeout 1024
+#define TinselLogMaxFlitsPerMsg 2
+#define TinselLogWordsPerMailbox 13
+#define TinselIntMultLatency 3
+#define TinselLogBytesPerMsg 6
+#define TinselLogFlitsPerMailbox 11
+#define TinselDCacheLogNumWays 4
+#define TinselLogNorthSouthLinks 0
+#define TinselDCachesPerDRAM 8
+#define TinselFPUsPerBoard 16
+#define TinselMeshXLenWithinBox 3
+#define TinselLogThreadsPerDRAM 9
+#define TinselMailboxMeshYBits 2
+#define TinselFetcherLogFlitBufferSize 5
+#define TinselLogThreadsPerMailbox 6
+#define TinselLogLinesPerDRAM 26
+#define TinselMeshYBits 3
+#define TinselLogWordsPerMsg 4
+#define TinselLogCoresPerFPU 2
+#define TinselSRAMLatency 8
+#define TinselLogBeatsPerDRAM 26
+#define TinselMeshYBits1 4
+#define TinselWordsPerFlit 4
+#define TinselDRAMBase 25165824
+#define TinselDeviceFamily '"Stratix V"'
+#define TinselSRAMLogMaxInFlight 5
+#define TinselBitsPerFlit 128
+#define TinselMeshYLenWithinBox 2
+#define TinselBoxMeshXLen 1
+#define TinselFetcherLogMsgsPerFlitBuffer 3
+#define TinselBoxMesh {{"ayres"}}
+#define TinselMaxFlitsPerMsg 4
+#define TinselBeatWidth 256
+#define TinselLogCoresPerMailbox 2
+#define TinselLogOffChipRAMBaseAddr 16
+#define TinselBeatsPerLine 1
+#define TinselPOLiteDRAMGlobalsLength 16384
+#define TinselLogBytesPerSRAMPartition 15
+#define TinselSharedInstrMem true
+#define TinselWordsPerLine 8
+#define TinselLogBytesPerDRAMPartition 21
+#define TinselNumNorthSouthLinks 1
+#define TinselFetcherMsgsPerFlitBuffer 8
+#define TinselBytesPerBeat 32
+#define TinselMaxBootImageBytes 576
+#define TinselLogCoresPerDCache 2
+#define TinselThreadsPerCore 16
+#define TinselBitsPerLine 256
+#define TinselThreadsPerBoard 1024
+#define TinselLogBytesPerMailbox 15
+#define TinselFPUOpMaxLatency 14
+#define TinselProgRouterMaxBurst 4
+#define TinselMacLatency 100
+#define TinselSRAMStoreLatency 2
+#define TinselLogBeatsPerLine 0
+#define TinselLogFPUsPerBoard 4
+#define TinselPOLiteProgRouterBase 25182208
+#define TinselBoxMeshYLen 1
+#define TinselMaxThreads 65536
+#define TinselDCacheSetsPerThread 4
+#define TinselLogRoutingEntryLen 5
+#define TinselSRAMBurstWidth 3
+#define TinselLogThreadsPerCore 4
+#define TinselMailboxMeshYLen 4
+#define TinselFPMultLatency 11
+#define TinselDRAMLatency 20
+#define TinselLogThreadsPerBoard 10
+#define TinselBoardsPerBox 7
+#define TinselBeatBurstWidth 3
+#define TinselLogInstrsPerCore 11
+#define TinselDCacheNumWays 16
+#define TinselDCacheLogSetsPerThread 2
+#define TinselLogDCachesPerDRAM 3
+#define TinselLogThreadsPerSRAM 8
+#define TinselMailboxesPerBoard 16
+#define TinselLogWordsPerFlit 2
+#define TinselLogTransmitBufferSize 10
+#define TinselDRAMLogMaxInFlight 5
+#define TinselMeshYBitsWithinBox 2
+#define TinselLogCoresPerBoard1 7
+#define TinselEnablePerfCount true
+#define TinselTransmitBound 20
+#define TinselDRAMsPerBoard 2
+#define TinselThreadsPerDRAM 512
+#define TinselLogMulticastBufferSize 9
+#define TinselLogBytesPerLine 5
+#define TinselCoresPerBoard 64
+#define TinselLogMailboxesPerBoard 4
+#define TinselSRAMsPerBoard 4
+#define TinselLogBytesPerSRAM 23
+#define TinselLogBytesPerBeat 5
+#define TinselFetcherLogIndQueueSize 1
+#define TinselSRAMAddrWidth 20
+#define TinselLogWordsPerBeat 3
+#define TinselThreadsPerMailbox 64
+#define TinselLogMsgsPerMailbox 9
+#define TinselFetchersPerProgRouter 8
+#define TinselPOLiteProgRouterLength 1048559616
+#define TinselLogMacRecvBufferSize 5
+#define TinselLogWordsPerLine 3
+#define TinselLogBeatsPerMem 27
+#define TinselLogBytesPerMem 32
+#define TinselMailboxMeshXBits 2
+#define TinselCoresPerMailbox 4
+#define TinselClockFreq 210
+#define TinselSRAMDataWidth 64
+#define TinselLogBytesPerSRAMBeat 3
+#define TinselLogMsgPtrQueueSize 6
+#define TinselDRAMGlobalsLength 1048576000
+#define TinselCoresPerDCache 4
+#define TinselUseCustomAccelerator false
+#define TinselMeshXBits 3
+#define TinselLogBeatsPerSRAM 18
+#define TinselLogThreadsPerFPU 6
+#endif
